Simplify project structure as we're now Java 8-only
