+++ gateway/engine/pom.xml
@@ -18,18 +18,7 @@
     <module>ispn</module>
     <module>es</module>
     <module>influxdb</module>
+    <module>vertx-eb-inmemory</module>
+    <module>prometheus</module>
   </modules>
-
-  <profiles>
-    <profile>
-      <id>java8</id>
-      <activation>
-        <jdk>[1.8,)</jdk>
-      </activation>
-      <modules>
-        <module>vertx-eb-inmemory</module>
-        <module>prometheus</module>
-      </modules>
-    </profile>
-  </profiles>
 </project>
