+++ gateway/platforms/pom.xml
@@ -17,18 +17,7 @@
     <module>war/micro</module>
     <module>war/tomcat8</module>
     <module>war/wildfly8</module>
+    <module>vertx3/vertx3</module>
+    <module>vertx3/vertx3-common</module>
   </modules>
-
-  <profiles>
-    <profile>
-      <id>java8</id>
-      <activation>
-        <jdk>[1.8,)</jdk>
-      </activation>
-      <modules>
-        <module>vertx3/vertx3</module>
-        <module>vertx3/vertx3-common</module>
-      </modules>
-    </profile>
-  </profiles>
 </project>
