+++ gateway/test/src/test/java/io/apiman/gateway/test/junit/vertx3/Vertx3GatewayTestServer.java
@@ -13,15 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.apiman.gateway.platforms.vertx3.junit.resttest;
-
-import java.io.File;
-import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
-import java.nio.file.Files;
-import java.util.concurrent.CountDownLatch;
-
-import org.codehaus.jackson.JsonNode;
+package io.apiman.gateway.test.junit.vertx3;
 
 import io.apiman.common.util.ReflectionUtils;
 import io.apiman.gateway.platforms.vertx3.common.config.VertxEngineConfig;
@@ -33,6 +25,14 @@ import io.vertx.core.DeploymentOptions;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
+
+import java.io.File;
+import java.io.IOException;
+import java.lang.reflect.InvocationTargetException;
+import java.nio.file.Files;
+import java.util.concurrent.CountDownLatch;
+
+import org.codehaus.jackson.JsonNode;
 
 /**
  * A Vert.x 3 version of the gateway test server
