+++ gateway/test/src/test/resources/test-configs/vertx3-es.json
@@ -1,5 +1,5 @@
 {
-    "factory" : "io.apiman.gateway.platforms.vertx3.junit.resttest.Vertx3GatewayTestServerFactory",
+    "factory" : "io.apiman.gateway.test.junit.vertx3.Vertx3GatewayTestServerFactory",
     "config" : "vertx3/conf-es.json",
-    "resetter" : "io.apiman.gateway.platforms.vertx3.junit.resttest.EsResetter"   
+    "resetter" : "io.apiman.gateway.test.junit.vertx3.EsResetter"   
 }
